refactor topic chips
